pyproject.toml
[project]
name = 'nih-crawl'
version = '0.1.0'
description = 'Exploratory topic dynamics and paradigm shift detection'
requires-python = '>=3.10'

dependencies = [
  'pandas',
  'numpy',
  'scikit-learn',
  'matplotlib',
  'tqdm',
  'pyarrow',
  'sentence-transformers',
  'python-dotenv',
]

[tool.setuptools]
package-dir = {'' = 'src'}

[tool.setuptools.packages.find]
where = ['src']

[project.optional-dependencies]
dev = [
  'jupyterlab', 
  'ipykernel',
]